backend/requirements.txt
Werkzeug==2.3.7
Flask==2.2.2
mysql-connector==2.2.9
flask-wtf == 1.2.1
flask-sqlalchemy == 3.1.1
python-dotenv == 1.0.0